+++ lib/Braintree/ApplePayCard.php
@@ -12,6 +12,7 @@ namespace Braintree;
  * @package    Braintree
  * @category   Resources
  *
+ * @property-read string $bin
  * @property-read string $cardType
  * @property-read string $createdAt
  * @property-read string $customerId