+++ lib/Braintree/PaymentMethodGateway.php
@@ -140,10 +140,20 @@ class PaymentMethodGateway
             'verificationAmount',
             ['paypal' => [
                 'payee_email',
+                'payeeEmail',
                 'order_id',
+                'orderId',
                 'custom_field',
+                'customField',
                 'description',
                 'amount',
+                ['shipping' =>
+                    [
+                        'firstName', 'lastName', 'company', 'countryName',
+                        'countryCodeAlpha2', 'countryCodeAlpha3', 'countryCodeNumeric',
+                        'extendedAddress', 'locality', 'postalCode', 'region',
+                        'streetAddress'],
+                ],
             ]],
         ];
         return [
@@ -164,7 +174,7 @@ class PaymentMethodGateway
 
     public static function createSignature()
     {
-        $signature = array_merge(self::baseSignature(), ['customerId']);
+        $signature = array_merge(self::baseSignature(), ['customerId', 'paypalRefreshToken', 'paypalVaultWithoutUpgrade']);
         return $signature;
     }