+++ lib/Braintree/WebhookTesting.php
@@ -77,6 +77,9 @@ class WebhookTesting
             case WebhookNotification::IDEAL_PAYMENT_FAILED:
                 $subjectXml = self::_idealPaymentFailedSampleXml($id);
                 break;
+            case WebhookNotification::GRANTED_PAYMENT_INSTRUMENT_UPDATE:
+                $subjectXml = self::_grantedPaymentInstrumentUpdateSampleXml();
+                break;
             default:
                 $subjectXml = self::_subscriptionSampleXml($id);
                 break;
@@ -441,6 +444,26 @@ class WebhookTesting
           <approval-url>https://example.com</approval-url>
           <ideal-transaction-id>1234567890</ideal-transaction-id>
         </ideal-payment>
+        ";
+    }
+
+    private static function _grantedPaymentInstrumentUpdateSampleXml()
+	{
+        return "
+		<granted-payment-instrument-update>
+		  <grant-owner-merchant-id>vczo7jqrpwrsi2px</grant-owner-merchant-id>
+		  <grant-recipient-merchant-id>cf0i8wgarszuy6hc</grant-recipient-merchant-id>
+		  <payment-method-nonce>
+			<nonce>ee257d98-de40-47e8-96b3-a6954ea7a9a4</nonce>
+			<consumed type='boolean'>false</consumed>
+			<locked type='boolean'>false</locked>
+		  </payment-method-nonce>
+		  <token>abc123z</token>
+		  <updated-fields type='array'>
+			<item>expiration-month</item>
+			<item>expiration-year</item>
+		  </updated-fields>
+		</granted-payment-instrument-update>
         ";
     }