+++ tests/integration/TransactionTest.php
@@ -736,6 +736,7 @@ class TransactionTest extends Setup
     {
         $result = Braintree\Transaction::sale([
             'amount' => '100.00',
+            'deviceSessionId' => 'abc123',
             'creditCard' => [
                 'cardholderName' => 'The Cardholder',
                 'number' => '5105105105105100',
@@ -746,6 +747,8 @@ class TransactionTest extends Setup
         $transaction = $result->transaction;
         $this->assertNotNull($transaction->riskData);
         $this->assertNotNull($transaction->riskData->decision);
+        $this->assertNull($transaction->riskData->id);
+        $this->assertNull($transaction->riskData->deviceDataCaptured);
     }
 
   public function testRecurring()
@@ -3466,7 +3469,7 @@ class TransactionTest extends Setup
 
         $this->assertTrue($result->success);
         $transaction = $result->transaction;
-        $this->assertEquals(Braintree\Transaction::SETTLING, $transaction->status);
+        $this->assertEquals(Braintree\Transaction::SETTLED, $transaction->status);
     }
 
   public function testCreate_withPayPalHandlesBadUnvalidatedNonces()
@@ -3705,7 +3708,7 @@ class TransactionTest extends Setup
   public function testIncludeProcessorSettlementResponseForSettlementDeclinedTransaction()
     {
         $result = Braintree\Transaction::sale([
-            "paymentMethodNonce" => Braintree\Test\Nonces::$paypalFuturePayment,
+            "paymentMethodNonce" => Braintree\Test\Nonces::$visaCheckoutVisa,
             "amount" => "100",
             "options" => [
                 "submitForSettlement" => true
@@ -3726,7 +3729,7 @@ class TransactionTest extends Setup
   public function testIncludeProcessorSettlementResponseForSettlementPendingTransaction()
     {
         $result = Braintree\Transaction::sale([
-            "paymentMethodNonce" => Braintree\Test\Nonces::$paypalFuturePayment,
+            "paymentMethodNonce" => Braintree\Test\Nonces::$visaCheckoutVisa,
             "amount" => "100",
             "options" => [
                 "submitForSettlement" => true
@@ -4236,8 +4239,7 @@ class TransactionTest extends Setup
         $this->assertEquals($result->transaction->billing["postalCode"], "95131");
     }
 
-
-    public function testTransactionsCanBeCreatedWithSharedParams()
+    public function testTransactionsCanBeCreatedWithSharedPaymentMethodToken()
     {
         $partnerMerchantGateway = new Braintree\Gateway([
             'environment' => 'development',
@@ -4283,6 +4285,71 @@ class TransactionTest extends Setup
         $result = $oauthAccesTokenGateway->transaction()->sale([
             'amount' => '100.00',
             'sharedPaymentMethodToken' => $creditCard->token,
+            'sharedCustomerId' => $customer->id,
+            'sharedShippingAddressId' => $address->id,
+            'sharedBillingAddressId' => $address->id
+        ]);
+
+        $this->assertEquals(
+            $result->transaction->shippingDetails->firstName,
+            $address->firstName
+        );
+        $this->assertEquals(
+            $result->transaction->billingDetails->firstName,
+            $address->firstName
+        );
+    }
+
+    public function testTransactionsCanBeCreatedWithSharedPaymentMethodNonce()
+    {
+        $partnerMerchantGateway = new Braintree\Gateway([
+            'environment' => 'development',
+            'merchantId' => 'integration_merchant_public_id',
+            'publicKey' => 'oauth_app_partner_user_public_key',
+            'privateKey' => 'oauth_app_partner_user_private_key'
+        ]);
+
+        $customer = $partnerMerchantGateway->customer()->create([
+            'firstName' => 'Joe',
+            'lastName' => 'Brown'
+        ])->customer;
+        $address = $partnerMerchantGateway->address()->create([
+            'customerId' => $customer->id,
+            'firstName' => 'Dan',
+            'lastName' => 'Smith',
+        ])->address;
+        $creditCard = $partnerMerchantGateway->creditCard()->create([
+            'customerId' => $customer->id,
+            'cardholderName' => 'Adam Davis',
+            'number' => '4111111111111111',
+            'expirationDate' => '05/2009'
+        ])->creditCard;
+
+        $oauthAppGateway = new Braintree\Gateway([
+            'clientId' =>  'client_id$development$integration_client_id',
+            'clientSecret' => 'client_secret$development$integration_client_secret'
+        ]);
+
+        $code = Test\Braintree\OAuthTestHelper::createGrant($oauthAppGateway, [
+            'merchant_public_id' => 'integration_merchant_id',
+            'scope' => 'read_write,shared_vault_transactions'
+        ]);
+
+        $credentials = $oauthAppGateway->oauth()->createTokenFromCode([
+            'code' => $code,
+        ]);
+
+        $oauthAccesTokenGateway = new Braintree\Gateway([
+            'accessToken' => $credentials->accessToken
+        ]);
+
+        $sharedNonce = $partnerMerchantGateway->paymentMethodNonce()->create(
+            $creditCard->token
+        )->paymentMethodNonce->nonce;
+
+        $result = $oauthAccesTokenGateway->transaction()->sale([
+            'amount' => '100.00',
+            'sharedPaymentMethodNonce' => $sharedNonce,
             'sharedCustomerId' => $customer->id,
             'sharedShippingAddressId' => $address->id,
             'sharedBillingAddressId' => $address->id